+++ gcp/handler.go
@@ -26,9 +26,10 @@ import (
 	"strings"
 )
 
-// Keys for [W3C Trace Context] attributes.
+// Keys for [W3C Trace Context] attributes by following [Trace Context in non-OTLP Log Formats].
 //
 // [W3C Trace Context]: https://www.w3.org/TR/trace-context/#traceparent-header-field-values
+// [Trace Context in non-OTLP Log Formats]: https://www.w3.org/TR/trace-context/#trace-id
 const (
 	// TraceKey is the key used by the [ID of the whole trace] forest and is used to uniquely
 	// identify a distributed trace through a system. It is represented as a 16-byte array,
@@ -36,18 +37,18 @@ const (
 	// All bytes as zero (00000000000000000000000000000000) is considered an invalid value.
 	//
 	// [ID of the whole trace]: https://www.w3.org/TR/trace-context/#trace-id
-	TraceKey = "trace-id"
+	TraceKey = "trace_id"
 	// SpanKey is the key used by the [ID of this request] as known by the caller.
 	// It is represented as an 8-byte array, for example, 00f067aa0ba902b7.
 	// All bytes as zero (0000000000000000) is considered an invalid value.
 	//
 	// [ID of this request]: https://www.w3.org/TR/trace-context/#parent-id
-	SpanKey = "span-id"
+	SpanKey = "span_id"
 	// TraceFlagsKey is the key used by an 8-bit field that controls [tracing flags]
 	// such as sampling, trace level, etc.
 	//
 	// [tracing flags]: https://www.w3.org/TR/trace-context/#trace-flags
-	TraceFlagsKey = "trace-flags"
+	TraceFlagsKey = "trace_flags"
 )
 
 // New creates a new Handler with the given Option(s).
@@ -206,7 +207,7 @@ func (h logHandler) Handle(ctx context.Context, record slog.Record) error { //no
 	if h.contextProvider != nil {
 		var found bool
 		record.Attrs(func(attr slog.Attr) bool {
-			if attr.Key == "trace-id" {
+			if attr.Key == TraceKey {
 				found = true
 
 				return false